Profile Management feature for bookstore API
This is a feature I implemented for a software engineering project where an API was created. I was in charge of creating a profile management feature that allows one to create a user, retrieve the list of users and their information, create a credit card and retrieve the data, and updating user information.